+++ core/record/_record/recordNodesUpdater.js
@@ -1,28 +1,24 @@
 import {
-  Dates,
   Objects,
-  Records,
   RecordUpdater as CoreRecordUpdater,
   RecordNodesUpdater as CoreRecordNodesUpdater,
-  RecordValidator,
   RecordUpdateResult,
   Promises,
-  Surveys,
 } from '@openforis/arena-core'
 
-import * as A from '@core/arena'
 import * as Survey from '@core/survey/survey'
 import * as NodeDef from '@core/survey/nodeDef'
 import * as Node from '@core/record/node'
-import * as Validation from '@core/validation/validation'
 import SystemError from '@core/systemError'
 
-import { NodeValues } from '../nodeValues'
 import * as RecordReader from './recordReader'
 import { NodeValueFormatter } from '../nodeValueFormatter'
+import { updateAttributeValue } from './recordNodeValueUpdater'
 
 const { createNodeAndDescendants, createRootEntity, deleteNodes } = CoreRecordUpdater
 const { updateNodesDependents } = CoreRecordNodesUpdater
+
+import { afterNodesUpdate } from './recordNodesUpdaterCommon'
 
 const _valueAdapterByType = {
   [NodeDef.nodeDefType.code]: ({ survey, record, parentNode, attributeDef, value }) => {
@@ -52,44 +48,6 @@ const _adaptValue = ({ survey, record, parentNode, attributeDef, value }) => {
   return valueAdapter ? valueAdapter({ survey, record, parentNode, attributeDef, value }) : value
 }
 
-const _updateAttributeValue = ({
-  survey,
-  record,
-  entity,
-  attributeDef,
-  attribute,
-  value,
-  dateModified: dataModifiedParam = null,
-  sideEffect = false,
-}) => {
-  if (
-    !NodeValues.isValueEqual({
-      survey,
-      nodeDef: attributeDef,
-      value: Node.getValue(attribute),
-      valueSearch: value,
-      record,
-      parentNode: entity,
-      strict: true,
-    }) ||
-    dataModifiedParam // always update attribute when dateModified changes
-  ) {
-    // update existing attribute (if value changed);
-    // do not update meta defaultValue applied flag (value could have been a default value already)
-    const attributeUpdated = A.pipe(
-      Node.assocValue(value),
-      Node.assocUpdated(true),
-      Node.assocDateModified(dataModifiedParam ?? new Date())
-    )(attribute)
-
-    const updateResult = new RecordUpdateResult({ record })
-    updateResult.addNode(attributeUpdated, { sideEffect })
-    return updateResult
-  }
-  // no updates performed
-  return null
-}
-
 const _addOrUpdateAttribute =
   ({ survey, entity, attributeDef, value: valueParam, sideEffect = false }) =>
   (record) => {
@@ -104,7 +62,7 @@ const _addOrUpdateAttribute =
       updateResult.addNode(attributeCreated, { sideEffect })
       return updateResult
     }
-    return _updateAttributeValue({ survey, record, entity, attributeDef, attribute, value, sideEffect })
+    return updateAttributeValue({ survey, record, entity, attributeDef, attribute, value, sideEffect })
   }
 
 const _addEntityAndKeyValues =
@@ -211,37 +169,6 @@ const _getOrCreateEntityByKeys =
     return { entity: entityInserted, updateResult }
   }
 
-const _afterNodesUpdate = async ({ survey, record, nodes, timezoneOffset, sideEffect = false }) => {
-  // output
-  const updateResult = new RecordUpdateResult({ record, nodes })
-
-  // 1. update dependent nodes
-  const updateResultDependents = updateNodesDependents({
-    survey,
-    record,
-    nodes,
-    timezoneOffset,
-    sideEffect,
-  })
-
-  updateResult.merge(updateResultDependents)
-
-  // 2. update node validations
-  const nodesValidation = await RecordValidator.validateNodes({
-    survey,
-    record: updateResult.record,
-    nodes: updateResult.nodes,
-  })
-  const recordValidationUpdated = A.pipe(
-    Validation.getValidation,
-    Validation.mergeValidation(nodesValidation),
-    Validation.updateCounts
-  )(updateResult.record)
-  updateResult.record = Validation.assocValidation(recordValidationUpdated)(updateResult.record)
-
-  return updateResult
-}
-
 const getOrCreateEntityByKeys =
   ({ survey, entityDefUuid, valuesByDefUuid, timezoneOffset, insertMissingNodes = false, sideEffect = false }) =>
   async (record) => {
@@ -258,7 +185,7 @@ const getOrCreateEntityByKeys =
     if (updateResultEntity) {
       updateResult.merge(updateResultEntity)
 
-      const dependentsUpdateResult = await _afterNodesUpdate({
+      const dependentsUpdateResult = await afterNodesUpdate({
         survey,
         record: updateResultEntity.record,
         nodes: updateResultEntity.nodes,
@@ -280,7 +207,7 @@ const updateAttributesInEntityWithValues =
 
       updateResult.merge(nodeUpdateResult)
 
-      const dependentsUpdateResult = await _afterNodesUpdate({
+      const dependentsUpdateResult = await afterNodesUpdate({
         survey,
         record: nodeUpdateResult.record,
         nodes: nodeUpdateResult.nodes,
@@ -362,295 +289,28 @@ const updateAttributesWithValues =
     return updateResult
   }
 
-const _findNodeWithSameUuid = (nodeSearch, nodesArray) =>
-  nodesArray.find((node) => Node.getUuid(node) === Node.getUuid(nodeSearch))
-
-const _getNodesArrayDifference = (nodes, otherNodes) => nodes.filter((node) => !_findNodeWithSameUuid(node, otherNodes))
-
-const _getNodesArrayIntersection = (nodes, otherNodes) =>
-  nodes.filter((node) => !!_findNodeWithSameUuid(node, otherNodes))
-
-const _replaceAttributeValueIfModifiedAfter = ({
-  survey,
-  attrDef,
-  recordTarget,
-  entityTarget,
-  attrSource,
-  attrTarget,
-  sideEffect,
-}) => {
-  const sourceDateModified = Node.getDateModified(attrSource)
-  const targetDateModified = Node.getDateModified(attrTarget)
-  if (Dates.isAfter(sourceDateModified, targetDateModified)) {
-    const attributeUpdateResult = _updateAttributeValue({
-      survey,
-      record: recordTarget,
-      entity: entityTarget,
-      attributeDef: attrDef,
-      attribute: attrTarget,
-      value: Node.getValue(attrSource),
-      dateModified: sourceDateModified,
-      sideEffect,
-    })
-    return attributeUpdateResult
-  }
-  return null
-}
-
-const _replaceUpdatedNodesInEntities = ({
-  survey,
-  recordSource,
-  recordTarget,
-  entitySource,
-  entityTarget,
-  sideEffect = false,
-}) => {
-  const includeAnalysis = false
-  const updateResult = new RecordUpdateResult({ record: recordTarget })
-
-  const entityDef = Survey.getNodeDefByUuid(Node.getNodeDefUuid(entitySource))(survey)
-
-  const metaSource = Node.getMeta(entitySource)
-  const metaTarget = Node.getMeta(entityTarget)
-  if (!Objects.isEqual(metaSource, metaTarget)) {
-    const entityTargetUpdated = A.pipe(Node.assocMeta(metaSource), Node.assocUpdated(true))(entityTarget)
-    updateResult.addNode(entityTargetUpdated)
-  }
-
-  Survey.getNodeDefChildren(
-    entityDef,
-    includeAnalysis
-  )(survey).forEach((childDef) => {
-    const childDefUuid = NodeDef.getUuid(childDef)
-
-    const childrenSource = RecordReader.getNodeChildrenByDefUuidUnsorted(entitySource, childDefUuid)(recordSource)
-    const childrenTarget = RecordReader.getNodeChildrenByDefUuidUnsorted(
-      entityTarget,
-      childDefUuid
-    )(updateResult.record)
-
-    // delete nodes that are not in source record
-    const childrenTargetToDelete = _getNodesArrayDifference(childrenTarget, childrenSource).map(Node.assocDeleted(true))
-    if (childrenTargetToDelete.length > 0) {
-      const childrenTargetToDeleteUuids = childrenTargetToDelete.map(Node.getUuid)
-      const nodesDeleteUpdateResult = Records.deleteNodes(childrenTargetToDeleteUuids, { sideEffect })(
-        updateResult.record
-      )
-      updateResult.merge(nodesDeleteUpdateResult)
-    }
-
-    // add new nodes (in source record but not in target record) to updateResult (and record)
-    _getNodesArrayDifference(childrenSource, childrenTarget).forEach((childSourceToAdd) => {
-      RecordReader.visitDescendantsAndSelf(childSourceToAdd, (visitedChildSource) => {
-        const newNodeToAdd = Node.assocCreated(true)(visitedChildSource) // new node for the server
-        delete newNodeToAdd[Node.keys.id] // clear internal id
-        updateResult.addNode(newNodeToAdd)
-      })(recordSource)
-    })
-
-    // update existing nodes (nodes in both source and target records)
-
-    _getNodesArrayIntersection(childrenSource, childrenTarget).forEach((childSource) => {
-      const childTargetToUpdate = _findNodeWithSameUuid(childSource, childrenTarget)
-      if (NodeDef.isAttribute(childDef)) {
-        const attrUpdateResult = _replaceAttributeValueIfModifiedAfter({
-          survey,
-          attrDef: childDef,
-          recordTarget: updateResult.record,
-          entityTarget,
-          attrSource: childSource,
-          attrTarget: childTargetToUpdate,
-          sideEffect,
-        })
-        if (attrUpdateResult) {
-          updateResult.merge(attrUpdateResult)
-        }
-      } else {
-        const childEntityUpdateResult = _replaceUpdatedNodesInEntities({
-          survey,
-          recordSource,
-          recordTarget: updateResult.record,
-          entitySource: childSource,
-          entityTarget: childTargetToUpdate,
-        })
-        updateResult.merge(childEntityUpdateResult)
-      }
-    })
-  })
-  return updateResult
-}
-
-const replaceUpdatedNodes =
-  ({ survey, recordSource, timezoneOffset, sideEffect = false }) =>
-  async (recordTarget) => {
-    const rootSource = RecordReader.getRootNode(recordSource)
-    const rootTarget = RecordReader.getRootNode(recordTarget)
-    if (Node.getUuid(rootTarget) !== Node.getUuid(rootSource)) {
-      // it should never happen...
-      throw new Error('error merging records: root entities have different uuids')
-    }
-    const updateResult = _replaceUpdatedNodesInEntities({
-      survey,
-      recordSource,
-      recordTarget,
-      entitySource: rootSource,
-      entityTarget: rootTarget,
-      sideEffect,
-    })
-    return _afterNodesUpdate({
-      survey,
-      record: updateResult.record,
-      nodes: updateResult.nodes,
-      timezoneOffset,
-      sideEffect,
-    })
-  }
-
-const findEntityByUuidOrKeys = ({ survey, record, entityDefUuid, parentEntity, uuid, keyValuesByDefUuid }) => {
-  const entityWithSameUuid = Records.getNodeByUuid(uuid)(record)
-  return (
-    entityWithSameUuid ??
-    Records.findEntityByKeyValues({ survey, record, parentEntity, entityDefUuid, keyValuesByDefUuid })
-  )
-}
-
-const mergeRecords =
-  ({ survey, recordSource, timezoneOffset, sideEffect = false }) =>
-  async (recordTarget) => {
-    const { cycle } = recordTarget
-    const rootSource = RecordReader.getRootNode(recordSource)
-    const rootTarget = RecordReader.getRootNode(recordTarget)
-
-    const updateResult = new RecordUpdateResult({ record: recordTarget })
-
-    const addNode = ({ node, parentUuid = undefined }) => {
-      const newNodeToAdd = Node.assocCreated(true)(node)
-      delete newNodeToAdd[Node.keys.id] // clear internal id
-      newNodeToAdd[Node.keys.recordUuid] = updateResult.record.uuid
-      if (parentUuid) {
-        newNodeToAdd[Node.keys.parentUuid] = parentUuid
-      }
-      updateResult.addNode(newNodeToAdd)
-    }
-
-    const stack = [{ entitySource: rootSource, entityTarget: rootTarget }]
-    while (stack.length > 0) {
-      const { entitySource, entityTarget } = stack.pop()
-
-      const entityDef = Surveys.getNodeDefByUuid({ survey, uuid: Node.getNodeDefUuid(entitySource) })
-      const childDefs = Surveys.getNodeDefChildrenSorted({ survey, nodeDef: entityDef, cycle })
-      childDefs.forEach((childDef) => {
-        const childDefUuid = NodeDef.getUuid(childDef)
-        const childrenSource = RecordReader.findNodeChildren(entitySource, childDefUuid)(recordSource)
-        const childrenTarget = RecordReader.findNodeChildren(entityTarget, childDefUuid)(recordTarget)
-        if (NodeDef.isSingle(childDef)) {
-          const childSource = childrenSource[0]
-          const childTarget = childrenTarget[0]
-          if (childSource && childTarget) {
-            if (NodeDef.isAttribute(childDef)) {
-              const attrUpdateResult = _replaceAttributeValueIfModifiedAfter({
-                survey,
-                attrDef: childDef,
-                recordTarget: updateResult.record,
-                entityTarget,
-                attrSource: childSource,
-                attrTarget: childTarget,
-                sideEffect,
-              })
-              if (attrUpdateResult) {
-                updateResult.merge(attrUpdateResult)
-              }
-            } else {
-              stack.push({ entitySource: childSource, entityTarget: childTarget })
-            }
-          }
-        } else if (NodeDef.isEntity(childDef)) {
-          childrenSource.forEach((childSource) => {
-            const keyValuesByDefUuid = Records.getEntityKeyValuesByDefUuid({
-              survey,
-              record: recordSource,
-              entity: childSource,
-            })
-            const childTarget = findEntityByUuidOrKeys({
-              survey,
-              record: recordTarget,
-              entityDefUuid: childDefUuid,
-              parentEntity: entityTarget,
-              uuid: Node.getUuid(childSource),
-              keyValuesByDefUuid,
-            })
-            if (childTarget) {
-              stack.push({ entitySource: childSource, entityTarget: childTarget })
-            } else {
-              RecordReader.visitDescendantsAndSelf(childSource, (visitedChildSource) => {
-                const parentUuid = visitedChildSource === childSource ? Node.getUuid(entityTarget) : undefined
-                addNode({ node: visitedChildSource, parentUuid })
-              })(recordSource)
-            }
-          })
-        } else {
-          // multiple attributes merge
-          const sourceValues = childrenSource.map(Node.getValue)
-          const targetValues = childrenTarget.map(Node.getValue)
-          if (
-            (sourceValues.length > 0 && sourceValues.length !== targetValues.length) ||
-            sourceValues.some((sourceValue, index) => {
-              const targetValue = childrenTarget[index]
-              return !NodeValues.isValueEqual({
-                survey,
-                nodeDef: childDef,
-                value: sourceValue,
-                valueSearch: targetValue,
-                record: updateResult.record,
-                parentNode: entityTarget,
-              })
-            })
-          ) {
-            // different values, replace all nodes
-            const childrenTargetToDeleteUuids = childrenTarget.map(Node.getUuid)
-            const nodesDeleteUpdateResult = Records.deleteNodes(childrenTargetToDeleteUuids, { sideEffect })(
-              updateResult.record
-            )
-            updateResult.merge(nodesDeleteUpdateResult)
-            childrenSource.forEach((childSource) => {
-              addNode({ node: childSource, parentUuid: Node.getUuid(entityTarget) })
-            })
-          }
-        }
-      })
-    }
-    return _afterNodesUpdate({
-      survey,
-      record: updateResult.record,
-      nodes: updateResult.nodes,
-      timezoneOffset,
-      sideEffect,
-    })
-  }
-
 const deleteNodesInEntityByNodeDefUuid =
   ({ survey, entity, nodeDefUuids, sideEffect = false }) =>
   async (record) => {
     const updateResult = new RecordUpdateResult({ record })
 
-    const nodeUudisToDelete = []
+    const nodeUuidsToDelete = []
     nodeDefUuids.forEach((nodeDefUuid) => {
       const children = RecordReader.getNodeChildrenByDefUuid(entity, nodeDefUuid)(record)
-      nodeUudisToDelete.push(...children.map(Node.getUuid))
+      nodeUuidsToDelete.push(...children.map(Node.getUuid))
     })
 
-    const nodesDeleteUpdateResult = await deleteNodes({ survey, record, nodeUuids: nodeUudisToDelete, sideEffect })
+    const nodesDeleteUpdateResult = await deleteNodes({ survey, record, nodeUuids: nodeUuidsToDelete, sideEffect })
     return updateResult.merge(nodesDeleteUpdateResult)
   }
 
 export const RecordNodesUpdater = {
+  afterNodesUpdate,
   createNodeAndDescendants,
   createRootEntity,
   getOrCreateEntityByKeys,
   updateNodesDependents,
   updateAttributesInEntityWithValues,
   updateAttributesWithValues,
-  replaceUpdatedNodes,
-  mergeRecords,
   deleteNodesInEntityByNodeDefUuid,
 }
