+++ core/record/_record/recordNodesUpdaterCommon.js
@@ -1,0 +1,41 @@
+import {
+  RecordValidator,
+  RecordUpdateResult,
+  RecordNodesUpdater as CoreRecordNodesUpdater,
+} from '@openforis/arena-core'
+
+import * as A from '@core/arena'
+import * as Validation from '@core/validation/validation'
+
+const { updateNodesDependents } = CoreRecordNodesUpdater
+
+export const afterNodesUpdate = async ({ survey, record, nodes, timezoneOffset, sideEffect = false }) => {
+  // output
+  const updateResult = new RecordUpdateResult({ record, nodes })
+
+  // 1. update dependent nodes
+  const updateResultDependents = updateNodesDependents({
+    survey,
+    record,
+    nodes,
+    timezoneOffset,
+    sideEffect,
+  })
+
+  updateResult.merge(updateResultDependents)
+
+  // 2. update node validations
+  const nodesValidation = await RecordValidator.validateNodes({
+    survey,
+    record: updateResult.record,
+    nodes: updateResult.nodes,
+  })
+  const recordValidationUpdated = A.pipe(
+    Validation.getValidation,
+    Validation.mergeValidation(nodesValidation),
+    Validation.updateCounts
+  )(updateResult.record)
+  updateResult.record = Validation.assocValidation(recordValidationUpdated)(updateResult.record)
+
+  return updateResult
+}
